+++ internal/kubernetes/vars.go
@@ -1,0 +1,10 @@
+package kubernetes
+
+var (
+	apps map[string]func(string) Application
+)
+
+const (
+	InstallKubectl = InstallType("kubectl")
+	InstallHelm    = InstallType("helm")
+)
